requirements.txt
pillow==9.3.0
pyyaml==6.0
tqdm~=4.64.1
python-dateutil~=2.8.2
requests~=2.31.0